requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dominate==2.5.2
Flask
Flask-Bootstrap
Flask-CKEditor
Flask-SQLAlchemy
Flask-WTF
idna==2.10
itsdangerous
Jinja2
MarkupSafe
requests==2.24.0
SQLAlchemy
urllib3==1.25.10
visitor==0.1.3
Werkzeug
WTForms
bleach
python-dotenv